Move MemoryEmailAddressContactModule to apps/memory module
- And MemoryFirebaseSubscriptionRepositoryModule
